// src/config/translations-interface.ts
export interface Translations {
    error: {
        characters: string;
        uppercase: string;
        lowercase: string;
        digit: string;
        character: string;
    };
    empty: {
        textContent: string;
    };
    titleOneStep: {
        textContent: string;
    };
    titleMultiStep: {
        step1: string;
        step2: string;
        step3: string;
    };
    popup: {
        textContent: string;
    };
    fields: {
        usernamePlaceholder: string;
        firstNamePlaceholder: string;
        lastNamePlaceholder: string;
        emailError: string;
        phonePlaceholder: string;
        diallingCode: string;
        phonError: string;
        passwordPlaceholder: string;
        passwordError: string;
        confirmPasswordPlaceholder: string;
        confirmPasswordError: string;
    };
    selects: {
        forFemale: string;
        forMale: string;
    };
    password: {
        textContent: string;
    };
    buttons: {
        valueSubmit: string;
        valuePrevious: string;
        valueNext: string;
        valueClose: string;
    };
}